Move `resolveEntity` into `settings.js`
Had to refer to `module.exports.`
because all of those functions get mocked in tests.
