+++ lib/controller.js
@@ -5,7 +5,6 @@ const logger = require('./util/logger');
 const settings = require('./util/settings');
 const zigbeeShepherdConverters = require('zigbee-shepherd-converters');
 const objectAssignDeep = require('object-assign-deep');
-const utils = require('./util/utils');
 
 // Extensions
 const ExtensionNetworkMap = require('./extension/networkMap');
@@ -233,7 +232,7 @@ class Controller {
     }
 
     publishEntityState(entityID, payload, cache) {
-        const entity = utils.resolveEntity(entityID);
+        const entity = settings.resolveEntity(entityID);
         const appSettings = settings.get();
         let messagePayload = {...payload};
