+++ lib/extension/deviceBind.js
@@ -1,6 +1,5 @@
 const settings = require('../util/settings');
 const logger = require('../util/logger');
-const utils = require('../util/utils');
 const Queue = require('queue');
 
 const topicRegex = new RegExp(`^${settings.get().mqtt.base_topic}/bridge/(bind|unbind)/.+$`);
@@ -55,7 +54,7 @@ class DeviceBind {
         }
 
         // Find source; can only be a device.
-        const sourceEntity = utils.resolveEntity(topic.ID);
+        const sourceEntity = settings.resolveEntity(topic.ID);
         const source = this.zigbee.getEndpoint(sourceEntity.ID);
 
         if (!source) {
@@ -64,7 +63,7 @@ class DeviceBind {
         }
 
         // Find target; can be a device or group.
-        const targetEntity = utils.resolveEntity(message.toString());
+        const targetEntity = settings.resolveEntity(message.toString());
         let target = null;
 
         if (targetEntity.type === 'device') {
