+++ lib/extension/devicePublish.js
@@ -2,7 +2,6 @@
 const settings = require('../util/settings');
 const zigbeeShepherdConverters = require('zigbee-shepherd-converters');
 const logger = require('../util/logger');
-const utils = require('../util/utils');
 
 const topicRegex = new RegExp(`^${settings.get().mqtt.base_topic}/.+/(set|get)$`);
 const postfixes = [
@@ -137,7 +136,7 @@ class DevicePublish {
         }
 
         // Resolve the entity
-        const entity = utils.resolveEntity(topic.ID);
+        const entity = settings.resolveEntity(topic.ID);
 
         // Get entity details
         let endpoint = null;
