+++ lib/extension/homeassistant.js
@@ -2,7 +2,6 @@ const zigbeeShepherdConverters = require('zigbee-shepherd-converters');
 const settings = require('../util/settings');
 const logger = require('../util/logger');
 const zigbee2mqttVersion = require('../../package.json').version;
-const utils = require('../util/utils');
 
 const configurations = {
     // Binary sensor
@@ -607,7 +606,7 @@ class HomeAssistant {
             return;
         }
 
-        const entity = utils.resolveEntity(entityID);
+        const entity = settings.resolveEntity(entityID);
         if (entity.type === 'device' && (!mapping[mappedModel.model] || !settings.getDevice(entity.ID))) {
             return;
         } else if (entity.type === 'group' && (!settings.getGroup(entity.ID))) {
