+++ lib/util/settings.js
@@ -204,6 +204,35 @@ function changeFriendlyName(old, new_) {
     return true;
 }
 
+// An entity can be either a group or a device.
+function resolveEntity(ID) {
+    let type = null;
+    let friendlyName = null;
+
+    if (module.exports.getIeeeAddrByFriendlyName(ID)) {
+        // Check if the ID is a friendly_name of a device.
+        friendlyName = ID;
+        ID = module.exports.getIeeeAddrByFriendlyName(ID);
+        type = 'device';
+    } else if (module.exports.getGroupIDByFriendlyName(ID)) {
+        // Check if the ID is a friendly_name of a group.
+        friendlyName = ID;
+        ID = Number(module.exports.getGroupIDByFriendlyName(ID));
+        type = 'group';
+    } else if (module.exports.getGroup(ID)) {
+        friendlyName = module.exports.getGroup(ID).friendly_name;
+        ID = Number(ID);
+        type = 'group';
+    } else {
+        // By default it is a device with ID as ID.
+        type = 'device';
+        const device = module.exports.getDevice(ID);
+        friendlyName = device ? device.friendly_name : ID;
+    }
+
+    return {ID: ID, type: type, friendlyName: friendlyName};
+}
+
 module.exports = {
     get: () => objectAssignDeep.noMutate(defaults, settings),
     write: () => write(),
@@ -219,6 +248,7 @@ module.exports = {
     getGroupIDByFriendlyName: (friendlyName) => getGroupIDByFriendlyName(friendlyName),
     changeFriendlyName: (old, new_) => changeFriendlyName(old, new_),
     changeDeviceOptions: (ieeeAddr, options) => changeDeviceOptions(ieeeAddr, options),
+    resolveEntity,
 
     addOnChangeHandler: (handler) => onChangeHandlers.push(handler),
